# apps/admin-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/oidc/callback": {
      "get": {
        "operationId": "OidcController_callback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Oidc"
        ]
      }
    },
...
  },
  "info": {
    "title": "AutoFlow Admin API",
    "description": "AutoCodeFlow Admin API — exported by test/openapi-export.e2e-spec.ts (ARCH-23). Single source of truth for generated frontend types (apps/admin-web gen:api-types) and the CI api-types-drift check. Regenerate with `npm run swagger:export` after any API contract change.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3105",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  },
  "security": [
    {
      "JWT": []
    }
  ]
}